Add multi-observatory averaging algorithm
Added Duff Stewart's multi-observatory averaging algorithm, using the DeltaFAlgorithm.py
module as a template. This averages per-timestamp values for a single channel, but multiple
observatories. A scale factor is allowed for each observatory. Simple checks are performed
to ensure observatory time series are compatible. Currently, it returns an ObsPy stream
with a channel name of 'MSD'. This probably needs to be changed.

Minor changes to related __init__.py files were made too.